Move mesh filename validator
