+++ flow360/__init__.py
@@ -6,9 +6,9 @@ from flow360.accounts_utils import Accounts
 from flow360.cli.api_set_func import configure_caller as configure
 from flow360.component.case import Case
 from flow360.component.geometry import Geometry
-from flow360.component.project import (
+from flow360.component.project import Project
+from flow360.component.project_utils import (
     GeometryFiles,
-    Project,
     SurfaceMeshFile,
     VolumeMeshFile,
 )
